% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia_svs_download.R
\name{tcia_svs_download}
\alias{tcia_svs_download}
\alias{tcia_svs_info}
\title{Download SVS Images from the Cancer Imaging Archive}
\usage{
tcia_svs_download(camic_ids, destdir = tempdir())

tcia_svs_info(camic_id)
}
\arguments{
\item{camic_ids}{\code{character()} A vector of CAMIC IDs obtained from the 'TCIA
Histopathology Custom Dataset Builder.json' file.}

\item{destdir}{\code{character(1)} The destination directory where the SVS images
will be saved. The default is \code{tempdir()}.}

\item{camic_id}{\code{character(1)} A single CAMIC ID.}
}
\value{
\itemize{
\item \code{tcia_svs_download}: The path to the downloaded file
\item \code{tcia_svs_info}: A list containing the metadata of the SVS image
}
}
\description{
This function downloads SVS images from the Cancer Imaging
Archive (TCIA) using the CAMIC ID obtained from the 'TCIA Histopathology
Custom Dataset Builder.json' file.
}
\details{
The json file can be obtained by navigating to the TCIA website
\url{https://www.cancerimagingarchive.net/} under 'Access The Data' > 'Search
Histopathology Portal' and clicking on the 'TCIA Histopathology Custom
Dataset Builder' link.
}
\examples{
if (interactive()) {
    tcia_svs_info("311781")
    tcia_svs_download("311781")
    tcia_svs_download(311781:311783)
}
}
