% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia_token_expires.R
\name{tcia_token_expires}
\alias{tcia_token_expires}
\title{Function to check when the API expires}
\usage{
tcia_token_expires()
}
\value{
`tcia_token_expires` returns a character string containing the expiration time of the access token.
}
\description{
`tcia_token_expires` : report the expiration time of the access token.
}
\examples{
tcia_token_expires()

}
