% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia_viewer.R
\name{tcia_viewer}
\alias{tcia_viewer}
\title{TCIA Viewer Widget}
\usage{
tcia_viewer(camic_id, width = "100\%", height = "600px", elementId = NULL)
}
\arguments{
\item{camic_id}{\code{character(1)} or \code{numeric(1)} The slide ID for caMicroscope
(e.g., "311781").}

\item{width, height}{\code{character(1)} A valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a string and
have \code{'px'} appended.}

\item{elementId}{\code{character(1)} (optional) An explicit ID for the widget's
root HTML element.}
}
\value{
An HTML widget object.
}
\description{
Embeds the caMicroscope mini viewer for a given slide ID using an iframe.
Requires an active internet connection.
}
\examples{
tcia_viewer("311781")

}
