% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcia_access_token.R
\name{tcia_access_token}
\alias{tcia_access_token}
\title{Set API and Obtain Access Token}
\usage{
tcia_access_token(
  username = "nbia_guest",
  client_id = "NBIA",
  grant_type = "password"
)
}
\arguments{
\item{username}{A character string for the username (default: "nbia_guest").}

\item{client_id}{A character string for the client ID (default: "NBIA").}

\item{grant_type}{A character string for the grant type (default:
"password").}
}
\value{
A character string containing the access token.
}
\description{
This function sends a request to the TCIA API to obtain an access token.
}
\examples{
tcia_access_token()

}
